+++ src/ila/event/minecraft/world/level/levelgen/structure/StructureEvent.cpp
@@ -1,0 +1,1 @@
+#include <ila/event/minecraft/world/level/levelgen/structure/StructureEvent.h>