+++ src/ila/include_all.h
@@ -42,6 +42,8 @@
 #include "ila/event/minecraft/world/level/block/MossGrowthEvent.h"
 #include "ila/event/minecraft/world/level/block/SculkCatalystAbsorbExperienceEvent.h"
 #include "ila/event/minecraft/world/level/block/SculkSpreadEvent.h"
+#include "ila/event/minecraft/world/level/levelgen/structure/StructureEvent.h"
+#include "ila/event/minecraft/world/level/levelgen/structure/VillageFeatureEvent.h"
 #include "ila/event/minecraft/world/level/LevelTickEvent.h"
 #include "ila/event/minecraft/world/level/WeatherUpdateEvent.h"
 #include "ila/event/minecraft/world/PistonPushEvent.h"