+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added VillageFeatureConstructionEvent @Lovelylavender4
+
 ### Changed
 
 - Refactored the namespaces of various events @Lovelylavender4